@@ -1,52 +1,41 @@
 {
-    "name": "smart-angular-new",
-    "version": "0.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "^6.0.6",
-        "@angular/common": "^6.0.6",
-        "@angular/compiler": "^6.0.6",
-        "@angular/core": "^6.0.6",
-        "@angular/forms": "^6.0.6",
-        "@angular/http": "^6.0.6",
-        "@angular/platform-browser": "^6.0.6",
-        "@angular/platform-browser-dynamic": "^6.0.6",
-        "@angular/router": "^6.0.6",
-        "@ng-bootstrap/ng-bootstrap": "^2.1.1",
-        "bootstrap": "^4.1.1",
-        "core-js": "^2.5.4",
-        "jquery": "~3.3.1",
-        "popper.js": "^1.14.3",
-        "rxjs": "^6.0.6",
-        "zone.js": "^0.8.26"
-    },
-    "devDependencies": {
-        "@angular/compiler-cli": "^6.0.6",
-        "@angular-devkit/build-angular": "~0.6.1",
-        "typescript": "~2.7.2",
-        "@angular/cli": "~6.0.1",
-        "@angular/language-service": "^6.0.6",
-        "@types/jasmine": "~2.8.6",
-        "@types/jasminewd2": "~2.0.3",
-        "@types/node": "~8.9.4",
-        "codelyzer": "~4.2.1",
-        "jasmine-core": "~2.99.1",
-        "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~1.7.1",
-        "karma-chrome-launcher": "~2.2.0",
-        "karma-coverage-istanbul-reporter": "~1.4.2",
-        "karma-jasmine": "~1.1.1",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "protractor": "~5.3.0",
-        "ts-node": "~5.0.1",
-        "tslint": "~5.9.1"
-    }
+  "name": "smart-angular",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "~13.2.0",
+    "@angular/cdk": "^13.2.0",
+    "@angular/common": "~13.2.0",
+    "@angular/compiler": "~13.2.0",
+    "@angular/core": "~13.2.0",
+    "@angular/forms": "~13.2.0",
+    "@angular/material": "^13.2.0",
+    "@angular/platform-browser": "~13.2.0",
+    "@angular/platform-browser-dynamic": "~13.2.0",
+    "@angular/router": "~13.2.0",
+    "rxjs": "~7.5.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.11.4"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~13.2.1",
+    "@angular/cli": "~13.2.1",
+    "@angular/compiler-cli": "~13.2.0",
+    "@types/jasmine": "~3.10.0",
+    "@types/node": "^12.11.1",
+    "jasmine-core": "~4.0.0",
+    "karma": "~6.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.1.0",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "typescript": "~4.5.2"
+  }
 }
